Move OVN bits to own area
Move OVN bits and fix unit tests due to changes in ops.testing
